+++ src/Environment.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace InvestecSdkPhp;
+
+enum Environment: string
+{
+    case SANDBOX = 'https://openapisandbox.investec.com/';
+    case PRODUCTION = 'https://openapi.investec.com/';
+}