+++ src/Requests/PrivateBanking/GetAccountsRequest.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace InvestecSdkPhp\Requests\PrivateBanking;
+
+use Saloon\Enums\Method;
+use Saloon\Http\Request;
+
+class GetAccountsRequest extends Request
+{
+    protected Method $method = Method::GET;
+
+    public function resolveEndpoint(): string
+    {
+        return '/za/pb/v1/accounts';
+    }
+}