+++ src/Resources/Resource.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace InvestecSdkPhp\Resources;
+
+use Saloon\Contracts\Connector;
+
+abstract class Resource
+{
+    public function __construct(protected Connector $connector)
+    {
+    }
+}