+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to `nikolaos-spyratos/investec-sdk-php` will be documented in this file.
+All notable changes to `nikspyratos/investec-sdk-php` will be documented in this file.
 
 ## [Unreleased]
 
@@ -8,3 +8,6 @@ All notable changes to `nikolaos-spyratos/investec-sdk-php` will be documented i
 - Base package details
 - Package quality enhancements: Pint, Phpstan, Rector
 - Testing: Pest
+- PrivateBanking requests + Resource (transfer/pay requests still TODO)
+- Individual (client credentials grant) connector
+- Business (OAuth2) connector (WIP)